remove irrelevant files from Src folder
